Add enter name screen

+++ FINAL/Core/Src/sensor.c
@@ -10,7 +10,7 @@
 uint16_t adc_receive[5];
 
 void sensor_init(){
-	HAL_ADC_Start_DMA(&hadc1, (uint16_t*)adc_receive, 5);
+	HAL_ADC_Start_DMA(&hadc1, (uint32_t*)adc_receive, 5);
 }
 
 void sensor_Read(){
